+++ CHANGES.rst
@@ -5,8 +5,9 @@ Changelog of lizard-map
 1.67 (unreleased)
 -----------------
 
+- Added custom OpenLayers._getScriptLocation.
+
 - Changed x-axis layout for graphs to have more ticks with no overlapping
-  
 
 - Jslint jquery.workspace.js.